refactor: move updatepub to lib/server/pub for organization and to prevent exposing unvalidated server action
