+++ core/lib/server/pub.ts
@@ -589,3 +589,121 @@ export const deletePub = async (pubId: PubsId) =>
 
 export const getPubStage = async (pubId: PubsId) =>
 	autoCache(db.selectFrom("PubsInStages").select("stageId").where("pubId", "=", pubId));
+
+export const updatePub = async ({
+	pubId,
+	pubValues,
+	stageId,
+	pubTypeId,
+	continueOnValidationError,
+}: {
+	pubId: PubsId;
+	pubValues: PubValues;
+	stageId?: StagesId;
+	pubTypeId?: PubTypesId;
+	continueOnValidationError: boolean;
+}) => {
+	const result = await db.transaction().execute(async (trx) => {
+		// Update the stage if a target stage was provided.
+		if (stageId !== undefined) {
+			await autoRevalidate(
+				trx.deleteFrom("PubsInStages").where("PubsInStages.pubId", "=", pubId)
+			).execute();
+			await autoRevalidate(
+				trx.insertInto("PubsInStages").values({ pubId, stageId })
+			).execute();
+		}
+
+		// Update the pub type if a pub type was provided.
+		if (pubId !== undefined) {
+			await autoRevalidate(
+				trx.updateTable("pubs").set({ pubTypeId }).where("id", "=", pubId)
+			).execute();
+		}
+
+		// First query for existing values so we know whether to insert or update.
+		// Also get the schemaName for validation. We want the fields that may not be in the pub, too.
+		const toBeUpdatedPubFieldSlugs = Object.keys(pubValues);
+
+		if (!toBeUpdatedPubFieldSlugs.length) {
+			return {
+				success: true,
+				report: "Pub updated successfully",
+			};
+		}
+
+		const existingPubFieldValues = await autoCache(
+			trx
+				.selectFrom("pub_fields")
+				.leftJoin("pub_values", (join) =>
+					join
+						.onRef("pub_values.fieldId", "=", "pub_fields.id")
+						.on("pub_values.pubId", "=", pubId)
+				)
+				.where("pub_fields.slug", "in", toBeUpdatedPubFieldSlugs)
+				.where("pub_fields.isRelation", "=", false)
+				.select([
+					"pub_values.id as pubValueId",
+					"pub_fields.slug",
+					"pub_fields.name",
+					"pub_fields.schemaName",
+				])
+		).execute();
+
+		const validationErrors = validatePubValuesBySchemaName({
+			fields: existingPubFieldValues,
+			values: pubValues,
+		});
+
+		const invalidValueSlugs = Object.keys(validationErrors);
+		if (invalidValueSlugs.length) {
+			if (continueOnValidationError) {
+				for (const slug of invalidValueSlugs) {
+					delete pubValues[slug];
+				}
+			} else {
+				throw new Error(Object.values(validationErrors).join(" "));
+			}
+		}
+
+		try {
+			// Insert, update on conflict
+			await autoRevalidate(
+				trx
+					.insertInto("pub_values")
+					.values((eb) =>
+						Object.entries(pubValues).map(([pubFieldSlug, pubValue]) => ({
+							id:
+								existingPubFieldValues.find(
+									(pubFieldValue) => pubFieldValue.slug === pubFieldSlug
+								)?.pubValueId ?? undefined,
+							pubId,
+							value: JSON.stringify(pubValue),
+							fieldId: eb
+								.selectFrom("pub_fields")
+								.where("pub_fields.slug", "=", pubFieldSlug)
+								.select("pub_fields.id"),
+						}))
+					)
+					.onConflict((oc) =>
+						oc.column("id").doUpdateSet((eb) => ({
+							value: eb.ref("excluded.value"),
+						}))
+					)
+			).execute();
+		} catch (error) {
+			return {
+				title: "Failed to update pub",
+				error: `${error.reason}`,
+				cause: error,
+			};
+		}
+
+		return {
+			success: true,
+			report: "Pub updated successfully",
+		};
+	});
+
+	return result;
+};
